Split microscope.filterwheel module into concrete and abstract.
* microscope/filterwheel.py: we have concrete device classes on their
  own modules, so move the new ThorlabsFilterWheel class into its own
  microscope.filterwheels module and FilterWheelBase to
  microscope.devices.
* microscope.filterwheels/thorlabs.py: ThorlabsFilterWheel moved here.
* microscope/devices.py: FilterWheelBase moved here.
* microscope/filterwheels/__init__.py: new microscope sub-package.
* microscope/testsuite/devices.py: fix import of FilterWheelBase.
* NEWS: mention this new classes.
* doc/supported-devices.rst: declare support for thorlabs filterwheel.
